requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==5.5
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.7
docopt==0.6.2
exceptiongroup==1.2.0
eyed3==0.9.7
fastapi==0.103.2
ffmpeg==1.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.8.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.39.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.5
mysql-connector-python==8.0.33
natsort==8.4.0
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pooch==1.8.0
prompt-toolkit==3.0.43
protobuf==3.20.3
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyAudioAnalysis==0.3.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
python-speech-features==0.6
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.27.0
streamlit==1.23.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.4
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
yarg==0.1.9
zipp==3.15.0
